ISSUE #86 * Create `minos.common.meta` module.
