+++ minos/common/meta.py
@@ -1,0 +1,50 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
+
+
+def classproperty(func):
+    """Decorator to use class properties"""
+    if not isinstance(func, (classmethod, staticmethod)):
+        func = classmethod(func)
+
+    return _ClassPropertyDescriptor(func)
+
+
+class _ClassPropertyDescriptor(object):
+    def __init__(self, fget, fset=None):
+        self.fget = fget
+        self.fset = fset
+
+    def __get__(self, obj, klass=None):
+        if klass is None:
+            klass = type(obj)
+        return self.fget.__get__(obj, klass)()
+
+    def __set__(self, obj, value):
+        if not self.fset:
+            raise AttributeError("can't set attribute")
+        type_ = type(obj)
+        return self.fset.__get__(obj, type_)(value)
+
+    def setter(self, func):
+        if not isinstance(func, (classmethod, staticmethod)):
+            func = classmethod(func)
+        self.fset = func
+        return self
+
+
+# noinspection PyPep8Naming
+class self_or_classmethod(classmethod):
+    """Decorator to use the same method as instance or as class based on the caller."""
+
+    # noinspection PyMethodOverriding
+    def __get__(self, instance, type_):
+        # noinspection PyUnresolvedReferences
+        get = super().__get__ if instance is None else self.__func__.__get__
+        # noinspection PyArgumentList
+        return get(instance, type_)
