+++ minos/common/model/aggregate.py
@@ -23,6 +23,9 @@ from ..exceptions import (
     MinosRepositoryManuallySetAggregateIdException,
     MinosRepositoryManuallySetAggregateVersionException,
 )
+from ..meta import (
+    self_or_classmethod,
+)
 from ..repository import (
     MinosRepository,
     MinosRepositoryAction,
@@ -30,18 +33,6 @@ from ..repository import (
 from .abc import (
     MinosModel,
 )
-
-
-# noinspection PyPep8Naming
-class class_or_instancemethod(classmethod):
-    """Decorator to use the same method as instance or as class based on the caller."""
-
-    # noinspection PyMethodOverriding
-    def __get__(self, instance, type_):
-        # noinspection PyUnresolvedReferences
-        get = super().__get__ if instance is None else self.__func__.__get__
-        # noinspection PyArgumentList
-        return get(instance, type_)
 
 
 class Aggregate(MinosModel):
@@ -53,7 +44,13 @@ class Aggregate(MinosModel):
     # FIXME: The ``broker`` attribute should be a reference to a ``MinosBaseBroker`` class instance.
     # noinspection PyShadowingBuiltins
     def __init__(
-        self, id: int, version: int, *args, _broker: str = None, _repository: MinosRepository = None, **kwargs,
+        self,
+        id: int,
+        version: int,
+        *args,
+        _broker: str = None,
+        _repository: MinosRepository = None,
+        **kwargs,
     ):
 
         super().__init__(id, version, *args, **kwargs)
@@ -133,7 +130,7 @@ class Aggregate(MinosModel):
         return instance
 
     # noinspection PyMethodParameters,PyShadowingBuiltins
-    @class_or_instancemethod
+    @self_or_classmethod
     async def update(self_or_cls, id: int = None, _repository: MinosRepository = None, **kwargs) -> Aggregate:
         """Update an existing ``Aggregate`` instance.
 
@@ -175,7 +172,7 @@ class Aggregate(MinosModel):
         self._fields |= new.fields
 
     # noinspection PyMethodParameters,PyShadowingBuiltins
-    @class_or_instancemethod
+    @self_or_classmethod
     async def delete(self_or_cls, id: Optional[int] = None, _repository: MinosRepository = None):
         """Delete the given aggregate instance.
